Use Discord's localised timestamps for dates
Some time ago, Discord introduced timestamps that automatically display a given UTC timestamp with seconds resolution in the client's own timezone. There are several options to display the timestamp.

This changes the previously static time and date displays in the info, uptime, and userinfo commands, and the heartbeat messages to these localised timestamps.

Also optimised the way the startup time is stored, ex-/imported, and used.

Removes pytz dependency.