Remove custom log level and use nextcord Logger
Previously, the log level was set to 25. With the removal of the custom level, it is now set to INFO (20).

The custom level is annoying to deal with, looks weird in use, and most likely not particularly good practice anyway.

We also never changed the logger to the nextcord one, meaning no nextcord logs were stored.