Migrate General cog to use Slash Commands
Request now uses a Form modal for the user to request a functionality.
Anime and Manga commands were moved here and migrated to Slash Commands.
Links and other config-based attributes are now stored as instance attributes.
The DTbot version is queried anew from the bot's config every time in case it has been refreshed in the meantime. This also avoids needing to reload the General cog whenever the load-related commands in the Dev cog trigger a change in the bot's config.