Migrate Dev cog to use Slash Commands
This is a GroupCog, meaning all its Slash Commands will start with /dev.
All commands in this cog are only available in the developer server.
In order to not have DTbot send its heartbeat messages, the '--no-heartbeat' argument must now be used. This replaces the old positional 1/0 argument.

Constants are now stored as instance attributes.

The (un)/(re)load commands can optionally sync the Developer Guild CommandTree, or the global CommandTree if the parameter is set to True by the invoking user.
These commands also provide an autofill which will show the relevant cogs (i.e. if RNG is not loaded, it will not be shown for unload/reload).

Updatever has been merged into updaterp, which now provides an option to reload the bot's config. Because the config is an attribute of the DTbot instance, all other cogs checking this attribute will always get the freshest version.
This also avoids having to reload the General cog, and for the bot to invoke its own Slash Command, which is not possible.

Refreshservers was moved here from the (retired) DatabaseManagement cog.