Further improve dancer listing in dance
Now removes the invoking user from the dancer list, meaning even when a user invokes the command and adds only themselves, the response still be the "dances alone" one.
Also changed the listing of two dancers to "dancer1 and dancer2" because the previous change made it "dancer1, and dancer2" which looks bad.