Add on_app_command_completion listener
This reintroduces the command usage stats previously in v2 and removed during v3 migration (due to a lack of such a listener). Functionally, it is the same as in v2, but without command_module data because the Command object doesn't have any way to tell which Cog it belongs from.