Remove commands that could violate Discovery ToS
We were notified by Discord that some (unspecified) content of DTbot was in violation of Discord's content policy for App Discovery.

In trying our best to adhere to these policies and to get DTbot accessible to more users, we have removed the following commands:
- Violent content: /bite, /choke, /kick, /kill, /whip
- Sexual content: /lewd
- Gambling-adjacent or addictive content: /roulette bet, /roulette spin, /roulette table, /russianroulette

We do not know if this will be the extent of removals, but to us, these seem to be the only commands that could possibly be violations.

We're sorry for the feature loss on v3, and hope you will continue to enjoy DTbot v3.