Several improvements to /whohas
- Added check for when a role has no members (used to cause DTbot to not respond)
- Increased the max. pages that will be displayed from 10 to 15.
- Reworked how the display is calculated and created, in order to avoid iterating over the full (potentially very large) role member list.

The new approach assumes that all role member mentions will be 19+5=24 characters long (19 is the current max size for a user ID, can be shorter for older accounts, 5 more characters to account for the mentioning string and listing separators), and divides the 2048 available characters by this value.
Up to 15 pages of Embeds will be created, and crucially, only as many members as can be displayed will be taken from the full member list (itertools.islice takes the first max_users mentions, or all if there are less than max_users role members).
The display hasn't changed, but now DTbot can all but guarantee an average of 85 users per page, for a max total of 1275 members total. (The first page has some additional text, meaning the real total will be a bit lower).