Fix PaginatorSession buttons not disappearing
Previously, a PaginatorSession would not remove the buttons from the interaction upon timing out, as was intended.
Also, it now uses the "emoji" kwarg instead of "label" to show the navigation emoji used on the NavButtons. This improves the rendering on mobile devices, as it will always use Discord's style of emoji instead of the device's standard set.

PaginatorSession doesn't inherit from discord.ui.View anymore, as it didn't need to do so in the first place. In its place, there is now a NavButtonView which handles a timeout by clearing the view of the related interaction (here: removes the NavButtons).

The previous timeout handling actually never applied to the view with the buttons in it, but only to the PaginatorSession object itself, as the buttons were held by another view object.