Use Enum for database procedures & refactor
Allows for better control over procedures and removing the hacky "returns" parameter of dbcallprocedure.

By introducing types of procedures (ones that return bool, int, or no values), introducing fitting function overloads which only accept BoolProcedures but also return a bool value, for example, was simple.

Allows for better extensibility in case of future types of procedures with other return types, or such that return a list of values.