Refactor logging setup
Extracted the setup of the stderr StreamHandler and the FileHandler to util.utils methods, which attach the configured handler to the Logger provided as a parameter.

Also introduced a 'in_dev_mode' property for better checking of the '--dev' flag in the launch arguments.

The private _file_handler attribute is initialised with the MISSING sentinel value from the discord.utils package to enable discord.py's default logging setup, if desired.