Extend /dev refreshservers
Now also invalidates (marks as left) servers the bot is no longer in.

This introduces new kinds of DBProcedure, namely GetServers and InvalidateMissingServer. GetServers returns a list of dictionaries, where the key is the field name and the value the field value. This makes GetServers a "returns many" procedure, for which an overload is introduced, as well as corresponding methods on the DBProcedure enum and a _ListProcedures type for procedures with more than a single result.