Initialise ConnectionPool in setup_hook instead
Moving the pool initialisation here enables the launcher.py script to properly ensure the used database exists. Previously, the pool would be created when the DTbot object is created in launcher.py, before the launcher could execute the CREATE DATABASE IF NOT EXISTS step, thus causing the pool to error if the database didn't exist yet.

Also removed the redundant super call for setup_hook, as it is blank by default.