Introduce python-dotenv for secrets storage
Moves storage of Database credentials, Developer command codes, and the bot token to a .env file instead of the config.ini used thus far.
Other bits may be moved to the .env file in the future.

Triggering the config reload in /dev updaterp will also reload the .env file, and override the values currently held in the environment.