Add Database migration capability to launcher
A new Config section contains migrations to be run during bot startup (if necessary; all migrations should include IF NOT EXISTS or appropriate equivalents).

Migrations are performed AFTER database and table creation, but BEFORE procedure creation to allow new procedures to reference a changed table schema without having to manually edit the database or table by hand.

This allows for database changes to be made automatically when a schema change is made to support new features.