Refactor Maths module with increased precision
Renamed parameters of `/maths add` to 'first' and 'second' for
consistency with the other Maths commands.

Increased rint precision to 3 digits.

The actual calculations are now done with the rinted values instead of
the raw inputs. This should avoid minor rounding inconsistencies.

Improved code readability by extracting rinting and calculations from
the f-strings.