Remove dead links from linklist.py
Some of the Imgur-hosted images have been removed for unknown reasons,
probably purging old and/or un(der)used images.

This commit removes all links confirmed as dead.

We will migrate the remaining images to our own hosting solution in
the near future to avoid relying on a host that deletes them randomly.