Remove py-rolldice and clean up /roll output
The output has been cleaned up a bit, removing the strikethrough of
dropped dice and replacing it with a separated string that is
prepended with a "Dropped" to let users more easily read the list of
dropped dice. For kept dice, "Kept" is only shown if there are also
dropped dice, otherwise the bot just shows the list of rolled numbers.

To remove the risk of typos in the options, they're now handled via an
enum whose member values are the strings shown to users during the use
of /roll.

We've been doing our own results calculation since introducing
py-rolldice due to issues with the xX flags. It was also used as not
much more than a glorified random number generator that happened to be
able to parse CritDice syntax, too.
For this reason, we have finally decided to remove the py-rolldice
library and replaced it with equivalent manual calculations.