Refactor DTbot class to have fewer side effects
All data previously assembled inside the DTbot class is now passed to
it from launcher.py.
This means the bot class no longer reads from the environment for the
database connection pool, causing it to no longer need to be set to
None with a type ignore in `__init__`.

The `in_dev_mode` property is now a private instance attribute because
it's not used anywhere else.

The run method override has been removed in favour of calling the
library's run method directly with the parameters from inside
launcher.py.

The logging utilities now behave as pure getters instead of modifying
a passed Logger directly.