Add pip-compile & repurpose requirements.txt
With the addition of the pyproject.toml file in 8c372ee, it is now
easily possible to compile a strictly defined list of dependencies,
all pinned to exact versions.

These changes mean that dependencies have been split into two files:

- requirements.txt:
  Production dependencies. These are needed for the bot to run.
- requirements-dev.txt:
  Development dependencies. These are needed for development work.

The pip-compile hooks only run when pyproject.toml is changed.